Add ostream stream operator to CharMap to allow printing to console. Add default constructor to StyledChar to conform to std::vector<StyledChar> requirements for resize() operation. Add bool to int casting in line in ConsoleColor implementation.

+++ ConsoleBlackjack/src/display/ConsoleColor.cpp
@@ -42,7 +42,7 @@ ValidColor CBJDisplay::ConsoleColor::colorByEnum() const
 {
     ValidColor color;
 
-    int colorCode = (m_red << 2) | (m_green << 1) | m_blue;
+    int colorCode = ((int)m_red << 2) | ((int)m_green << 1) | (int)m_blue;
 
     switch (colorCode)
     {
